requirements.txt
Django==4.0
psycopg2-binary==2.8.6
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-xml==2.0.0
PyJWT==2.3.0

